ISSUE #5257 - move files
