+++ frontend/src/v5/ui/components/viewer/cards/cardFilters/filtersAccordion/filtersAccordion.component.tsx
